requirements.txt
discord.py>=1.2.5
flask
aiohttp
asyncio
python-dotenv
